backend/requirements.txt
asgiref==3.2.7
autopep8==1.5.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
cymem==2.0.3
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
en-core-web-sm==2.2.5
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
murmurhash==1.0.2
numpy==1.18.2
plac==1.1.3
preshed==3.0.2
psycopg2==2.8.5
pycodestyle==2.5.0
pytz==2019.3
requests==2.23.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.0
tqdm==4.45.0
urllib3==1.25.8
wasabi==0.6.0
whitenoise==5.0.1
zipp==3.1.0
